ok rangemetn et creeation d'un squelette pour les thread et ode manuel fonctionelle
